請求額コンポーネントの準備
