+++ packages/kokoas-client/src/pages/projInvoice/FormInvoice.tsx
@@ -11,14 +11,11 @@ import { paymentLabels } from '../projContracts';
 import { FormikSelect } from '../../components/ui/selects';
 import { PlannedPaymentDate } from './fieldComponents/PlannedPaymentDate';
 import { useResolveParams } from './hooks/useResolveParams';
-import { FormikMoneyField } from 'kokoas-client/src/components/ui/textfield/FormikMoneyField';
-import { useContractAmount } from './hooks/useContractAmount';
-import { useState } from 'react';
-import { debounce } from 'lodash';
 import { SearchProjects } from 'kokoas-client/src/components/ui/textfield';
 import { useNavigate } from 'react-router-dom';
 import { generateParams } from 'kokoas-client/src/helpers/url';
 import { pages } from '../Router';
+import { BillingAmount } from './fieldComponents/BillingAmount';
 
 
 
@@ -27,13 +24,9 @@ export const FormInvoice = () => {
   const navigate = useNavigate();
 
   const { projId, projName } = values;
-  const [billingAmount, setBillingAmount] = useState(useContractAmount(projId));
 
   useResolveParams();
 
-  const handleChange = (e: any) => {
-    debounce(() => { setBillingAmount(e.target.value); }, 2000);
-  };
 
 
   return (
@@ -87,12 +80,7 @@ export const FormInvoice = () => {
 
         {/* 請求金額 */}
         <Grid item xs={12} md={6}>
-          <FormikMoneyField
-            label='請求額'
-            name={getFieldName('billingAmount')}
-            onChange={handleChange}
-            value={billingAmount}
-          />
+          <BillingAmount projId={projId} />
         </Grid>
         <Grid item md={6} />
